Yesod.Form.Jquery

Description

Some fields spiced up with jQuery UI.

Documentation

class YesodJquery a where

Methods

urlJqueryJs :: a -> Either (Route a) Text

The jQuery Javascript file. Note that in upgrades to this library, the version of jQuery referenced, or where it is downloaded from, may be changed without warning. If you are relying on a specific version of jQuery, you should give an explicit URL instead of relying on the default value.

Currently, the default value is jQuery 1.7 from Google's CDN.

urlJqueryUiJs :: a -> Either (Route a) Text

The jQuery UI 1.8 Javascript file.

urlJqueryUiCss :: a -> Either (Route a) Text

The jQuery UI 1.8 CSS file; defaults to cupertino theme.

urlJqueryUiDateTimePicker :: a -> Either (Route a) Text

jQuery UI time picker add-on.

jqueryDayField :: (RenderMessage master FormMessage, YesodJquery master) => JqueryDaySettings -> Field sub master Day

jqueryAutocompleteField :: (RenderMessage master FormMessage, YesodJquery master) => Route master -> Field sub master Text

googleHostedJqueryUiCss :: Text -> Text

Gets the Google hosted jQuery UI 1.8 CSS file with the given theme.

data JqueryDaySettings

Constructors

JqueryDaySettings

Fields

jdsChangeMonth :: Bool

jdsChangeYear :: Bool

jdsYearRange :: String

jdsNumberOfMonths :: Either Int (Int, Int)

class Default a where

A class for types with a default value.

Methods

def :: a

The default value for this type.