Yesod.Form.Nic

Description

Provide the user with a rich text editor.

According to NIC editor homepage it is not actively maintained since June 2012. There is another better alternative — open sourced Summernote editor released under MIT licence. You can use Summernote in your Yesod forms via separately distributed yesod-form-richtext package.

Documentation

class Yesod a => YesodNic a where

Minimal complete definition

Nothing

Methods

urlNicEdit :: a -> Either (Route a) Text

NIC Editor Javascript file.

nicHtmlField :: YesodNic site => Field (HandlerFor site) Html