Yesod.Form.Types

Helpers

data Enctype

The encoding type required by a form. The ToHtml instance produces values that can be inserted directly into HTML.

Constructors

UrlEncoded

Multipart

data FormResult a

A form can produce three different results: there was no data available, the data was invalid, or there was a successful parse.

The Applicative instance will concatenate the failure messages in two FormResults.

Constructors

FormMissing

FormFailure [Text]

FormSuccess a

data FormMessage

Constructors

MsgInvalidInteger Text

MsgInvalidNumber Text

MsgInvalidEntry Text

MsgInvalidUrl Text

MsgInvalidEmail Text

MsgInvalidTimeFormat

MsgInvalidHour Text

MsgInvalidMinute Text

MsgInvalidSecond Text

MsgInvalidDay

MsgCsrfWarning

MsgValueRequired

MsgInputNotFound Text

MsgSelectNone

MsgInvalidBool Text

MsgBoolYes

MsgBoolNo

MsgDelete

type Env = Map Text [Text]

type FileEnv = Map Text [FileInfo]

data Ints

Constructors

IntCons Int Ints

IntSingle Int

Form

type MForm m a = RWST (Maybe (Env, FileEnv), HandlerSite m, [Lang]) Enctype Ints m a

newtype AForm m a

Constructors

AForm

Fields

unAForm :: (HandlerSite m, [Text]) -> Maybe (Env, FileEnv) -> Ints -> m (FormResult a, [FieldView (HandlerSite m)] -> [FieldView (HandlerSite m)], Ints, Enctype)

Build forms

data Field m a

Constructors

Field

Fields

fieldParse :: [Text] -> [FileInfo] -> m (Either (SomeMessage (HandlerSite m)) (Maybe a))

fieldView :: FieldViewFunc m a

fieldEnctype :: Enctype

data FieldSettings master

Constructors

FieldSettings

Fields

fsLabel :: SomeMessage master

fsTooltip :: Maybe (SomeMessage master)

fsId :: Maybe Text

fsName :: Maybe Text

fsAttrs :: [(Text, Text)]

data FieldView site

Constructors

FieldView

Fields

fvLabel :: Markup

fvTooltip :: Maybe Markup

fvId :: Text

fvInput :: WidgetT site IO ()

fvErrors :: Maybe Markup

fvRequired :: Bool

type FieldViewFunc m a

Arguments

= Text
ID

-> Text
Name

-> [(Text, Text)]
Attributes

-> Either Text a
Either (invalid text) or (legitimate result)

-> Bool
Required?

-> WidgetT (HandlerSite m) IO ()