Yesod.Form.Types

Helpers

data Enctype

The encoding type required by a form. The ToHtml instance produces values that can be inserted directly into HTML.

Constructors

UrlEncoded

Multipart

data FormResult a

A form can produce three different results: there was no data available, the data was invalid, or there was a successful parse.

The Applicative instance will concatenate the failure messages in two FormResults. The Alternative instance will choose FormFailure before FormSuccess, and FormMissing last of all.

Constructors

FormMissing

FormFailure [Text]

FormSuccess a

data FormMessage

Constructors

MsgInvalidInteger Text

MsgInvalidNumber Text

MsgInvalidEntry Text

MsgInvalidUrl Text

MsgInvalidEmail Text

MsgInvalidTimeFormat

MsgInvalidHour Text

MsgInvalidMinute Text

MsgInvalidSecond Text

MsgInvalidDay

MsgCsrfWarning

MsgValueRequired

MsgInputNotFound Text

MsgSelectNone

MsgInvalidBool Text

MsgBoolYes

MsgBoolNo

MsgDelete

type Env = Map Text [Text]

type FileEnv = Map Text [FileInfo]

data Ints

Constructors

IntCons Int Ints

IntSingle Int

Form

type WForm m a = MForm (WriterT [FieldView (HandlerSite m)] m) a

MForm variant stacking a WriterT. The following code example using a monadic form MForm:

formToAForm $ do
  (field1F, field1V) <- mreq textField MsgField1 Nothing
  (field2F, field2V) <- mreq (checkWith field1F textField) MsgField2 Nothing
  (field3F, field3V) <- mreq (checkWith field1F textField) MsgField3 Nothing
  return
    ( MyForm <$> field1F <*> field2F <*> field3F
    , [field1V, field2V, field3V]
    )

Could be rewritten as follows using WForm:

wFormToAForm $ do
  field1F <- wreq textField MsgField1 Nothing
  field2F <- wreq (checkWith field1F textField) MsgField2 Nothing
  field3F <- wreq (checkWith field1F textField) MsgField3 Nothing
  return $ MyForm <$> field1F <*> field2F <*> field3F

Since: 1.4.14

type MForm m a = RWST (Maybe (Env, FileEnv), HandlerSite m, [Lang]) Enctype Ints m a

newtype AForm m a

Constructors

AForm

Fields

unAForm :: (HandlerSite m, [Text]) -> Maybe (Env, FileEnv) -> Ints -> m (FormResult a, [FieldView (HandlerSite m)] -> [FieldView (HandlerSite m)], Ints, Enctype)

Build forms

data Field m a

Constructors

Field

Fields

fieldParse :: [Text] -> [FileInfo] -> m (Either (SomeMessage (HandlerSite m)) (Maybe a))

fieldView :: FieldViewFunc m a

fieldEnctype :: Enctype

data FieldSettings master

Constructors

FieldSettings

Fields

fsLabel :: SomeMessage master

fsTooltip :: Maybe (SomeMessage master)

fsId :: Maybe Text

fsName :: Maybe Text

fsAttrs :: [(Text, Text)]

data FieldView site

Constructors

FieldView

Fields

fvLabel :: Markup

fvTooltip :: Maybe Markup

fvId :: Text

fvInput :: WidgetFor site ()

fvErrors :: Maybe Markup

fvRequired :: Bool

type FieldViewFunc m a

Arguments

= Text
ID

-> Text
Name

-> [(Text, Text)]
Attributes

-> Either Text a
Either (invalid text) or (legitimate result)

-> Bool
Required?

-> WidgetFor (HandlerSite m) ()